Fix: className in CustomMessage, avoid render undefined in ScrollButton and BannerUnreadMessages (#2695)
## Description
This PR fixes two issues introduced in version 0.24.0 of @botonic/react.

1) Custom components did not have the customMessage name as className.

2) The ScrollButton and BannerUnreadMessages components rendered nothing
if this feature was not used.

## Context
1) Custom components did not have the customMessage name as className.
This happened because the deserialize function of the Custom component
did not pass the new sentBy and isUnread attributes.
As these attributes were not serialised in the Message component, a
check was made to define the className that verified that the
CustomMessage was sent by the bot. Prior to version 0.24.0 if from
(sentBy now) was not defined then SENDERS.bot was set by default.

2) The ScrollButton and BannerUnreadMessages components rendered nothing
if this feature was not used.
If the feature is enabled by default or using custom components to
display these components the error did not occur but if the feature was
not enabled the webchat would break. Now when this feature is not active
it returns null.