Update webchat user using  WebchatSettings component (#2696)
## Description
Be able to update the webchat user from a bot action using
WebchatSettings component.

## Context
We have always had a problem with the synchronisation of the
session.user of the webchat (front-end) and the session.user of the bot
(lambda).
In each input that is sent from the webchat the session is attached to
make a merge of this in the backend (hubtype-backend) and this updates
it to the bot's lambda.
This causes that if you update the session in the bot's lambda action,
in the next action you have the session with the frontend values.
For example: 
- When you start a webchat you normally give a value of
user.extra_data.language to indicate in which language the bot is going
to respond. This value can be obtained and updated with the language in
which the web is being viewed.
- Then in an action that updates the language through a click on a
button changing this value in the session.user.extra_data.language of
the lambda, but in the next action this value returned to have the value
it had in the webchat. To solve this you could create a custom message
that updated this value first in the front-end but this was not the best
solution.

## Approach taken / Explain the design
To update the session.user of the webchat from a bot action, now you can
use the WebchatSettings component, to which you can pass the user that
you want to update in the webchat (front-end).
For now I will only allow the extra_data to be updated as there may be
sensitive data in the lambda session.

## To document / Usage example

SetLanguageAction -> which updates the language in the bot's lambda, In
the render we pass the WebchatSettings with the new values and the
content we want to render (in this case using FlowBuilder)

```typescript
import {
  FlowBuilderActionProps,
  FlowBuilderMultichannelAction,
} from '@botonic/plugin-flow-builder'
import { RequestContext, WebchatSettings } from '@botonic/react'

interface SetLanguageActionProps extends FlowBuilderActionProps {
  user: { extra_data: any }
}

export class SetLanguageAction extends React.Component<SetLanguageActionProps> {
  static contextType = RequestContext
  static async botonicInit(
    request: BotRequest
  ): Promise<SetLanguageActionProps> {
    const { cmsPlugin, payload } = getRequestData(request)

    const userData = request.session.user.extra_data
    const language = payload.split(SEPARATOR)[1] as string
    request.session.user.extra_data.language = language
    const user = {
      extra_data: {
        language: userData.language,
        country: userData.country,
      },
    }

    return {
      contents,
      user,
    }
  }

  render() {
    return (
      <>
        <WebchatSettings user={this.props.user} />
        <FlowBuilderMultichannelAction contents={this.props.contents} />
      </>
    )
  }
}
```
<!--
- How this is used?
- If possible, provide a snippet of code with a usage example.
-->

## Testing

The pull request...

- has unit tests
- has integration tests
- doesn't need tests because... **[provide a description]**