plugin-flow-builder: add the go-to-flow type node (#2737)
## Description
Add the go-to-flow node. 
This allows the use of splitting the flow into smaller flows. 

## Context
When an id is received that is from a go-to-flow node, the array of
flows is searched for the start node of that flow and the content node
that starts the flow is returned, instead of the go-to-flow node.


## To document / Usage example
Inside the json containing all the bot's flow we will have go-to-flow
nodes and an array with the information of each flow

This is a go-to-flow node
 ```
 {
  "id": "e3aa7335-b96f-4179-bc76-00fd00f4e4e1",
  "code": "",
  "meta": {
    "x": 930.8903465091073,
    "y": -533.8866771888627
  },
  "follow_up": null,
  "target": null,
  "flow_id": "8d527e7d-ea6d-5422-b810-5b4c8be7657b",
  "type": "go-to-flow",
  "content": {
    "flow_id": "7e7bf40e-756e-4011-866e-ffc43e60de92"
  }
}
```

And this is a flows array
```
flows: [
  {
    "id": "8d527e7d-ea6d-5422-b810-5b4c8be7657b",
    "name": "Main",
    "start_node_id": "71e7f609-57e1-4bc1-9545-e2de6e3bc9a4"
  },
  {
    "id": "03bafba6-c0fa-5449-9d42-bd98b44fe370",
    "name": "Fallback",
    "start_node_id": "f3931bce-7de3-5c7a-8287-81f0292ee4f3"
  },
  {
    "id": "7e7bf40e-756e-4011-866e-ffc43e60de92",
    "name": "Hacer un pedido",
    "start_node_id": "9b707d59-83f6-4669-b2a8-5e0595781735"
  },
  {
    "id": "dd173adc-3efc-4ba9-be5f-73fb903deb71",
    "name": "hacer una devolucion",
    "start_node_id": "b05351fa-1ec5-47c5-8653-286c0cf2eafe"
  }
]
```