plugin-flow-builder: handle bot action on the flow builder plugin #BLT-483 (#2769)
## Description

- adds the functionality to use BotActionNode created in the FlowBuilder
frontend.
- add a function to getPayloadParams in JSON format when receiving a
payload.
- add a base type for payloadParams 

## Context

A BotAction node can target 4 types of nodes: Button, Handoff, Keyword
and Intent

When the target is a Button (from Text, Carousel or WhatsappButtonList)
or a Handoff in the botonicInit function of the plugin-flow-builder we
must obtain the payload defined in the BotAction node and replace it.
This way when the user clicks on a button or executes a handoff the
payload will be the correct one.

With Keywords and Intents it is different, what we want is that when the
bot executes the routes, the payload is already in the request so that
the route of the action indicated by the payload is executed. To do this
we have to make the logic of building the payload in the pre function of
the plugin-flow-builder. This function is executed before the routes.
**IMPORTANT -> For the case where an intent points directly to a nod
BotAction to work, the plugin-hubtype-babel must be placed before the
plugin-flow-builder.**

## To document / Usage example

Example of how a bot action would be executed by receiving a payload
defined with a BotActionNode.

Here we use the function getPaylodParams, one of the parameters will be
followUpId if the BotAcitonNode has a follow up defined. With this Id we
retrieve the contents for rendering.

```typescript
import {
  FlowBuilderActionProps,
  FlowBuilderMultichannelAction,
} from '@botonic/plugin-flow-builder'
import { PayloadParamsBase } from '@botonic/plugin-flow-builder/lib/esm'
import { RequestContext } from '@botonic/react'
import React from 'react'

import { BotRequest } from '../types'
import { getRequestData } from '../utils/actions'

export class TestAction extends FlowBuilderMultichannelAction {
  static contextType = RequestContext
  static async botonicInit(
    request: BotRequest
  ): Promise<FlowBuilderActionProps> {
    const { cmsPlugin, botContext, payload } = getRequestData(request)
    const payloadParams = cmsPlugin.getPayloadParams<PayloadParamsBase>(payload)

    console.log('Do some logic with payloadParams', { payloadParams })

    const contents = await cmsPlugin.getContentsById(
      payloadParams.followUpId,
      botContext.locale
    )
    return { contents }
  }
}
```

## Testing

The pull request...

- has unit tests
- has integration tests
- doesn't need tests because... **[provide a description]**