plugin-flow-builder: add custom conditional (#2791)
## Description

Add the function to resolve custom conditional node. 
This allows to check any variable(string, boolean, number) inside
`request.session.user.extra_data`

## Context

These conditionals have one output for each check plus a `default`. 
When the variable being tested is a boolean the false and default output
is the same. This resolves the case where you are testing a boolean that
is currently undefined.

You can use variables that are nested in objects within extra_data. 
For example if you have the following extra_data object:
```typescript
const extra_data = {
  isLoggedIn: true,
  numberOfBags: 1,
  vehicle: 'bike',
  booking: {
    payment: {
      isPaid: false,
      type: 'cash',
    },
  },
}
```
You can access any variable by specifying the name with a string. If
nested as isPaid or type the string must be dotted as if you were
accessing the object in the code editor.

If the string does not start with either session or input, this variable
will be looked for within `session.user.extra_data` (most of the
variables that will be used are here). For example

- isLoggedIn
- numberOfBags
- vehicle 
- booking.payment.isPaid,
- booking.payment.type

If the string starts with session or input we will look for this
variable in `request`. For example
- session.user.name 
- input.data

If the value to replace in a text message is the `_access_token`, it
will not be displayed.

## Approach taken / Explain the design

example of the JSON received from a custom conditional that checks a
string variable called vehicle inside extra_data

```typescript
{
      id: 'bb34a3be-9cea-41e7-9616-4604afe00107',
      code: 'CHECK_VEHICLE',
      meta: {
        x: 242.24137237222322,
        y: 316.4777522774672,
      },
      follow_up: null,
      target: null,
      flow_id: '8d527e7d-ea6d-5422-b810-5b4c8be7657b',
      type: 'function',
      content: {
        action: 'check-bot-variable',
        arguments: [
          {
            type: 'string',
            name: 'keyPath',
            value: 'vehicle',
          },
        ],
        result_mapping: [
          {
            result: 'car',
            target: {
              id: '5fb5360e-3d9e-4dea-bdff-51e277b67dcb',
              type: 'text',
            },
          },
          {
            result: 'bike',
            target: {
              id: '63d151eb-993a-4e80-9132-8107e9b928d1',
              type: 'text',
            },
          },
          {
            result: 'default',
            target: {
              id: '8f81e367-6661-4be0-835e-7d598a3c9562',
              type: 'text',
            },
          },
        ],
      },
    },
```
example of JSON recived from a custom conditional that checks a string
variable called type inside extra_data.payment.type
```typescript
{
      id: 'bb34a3be-9cea-41e7-9616-4604afe00115',
      code: 'CHECK_PAYMENT_TYPE',
      meta: {
        x: 242.24137237222322,
        y: 316.4777522774672,
      },
      follow_up: null,
      target: null,
      flow_id: '8d527e7d-ea6d-5422-b810-5b4c8be7657b',
      type: 'function',
      content: {
        action: 'check-bot-variable',
        arguments: [
          {
            type: 'string',
            name: 'keyPath',
            value: 'booking.payment.type',
          },
        ],
        result_mapping: [
          {
            result: 'card',
            target: {
              id: '5fb5360e-3d9e-4dea-bdff-51e277b67d00',
              type: 'text',
            },
          },
          {
            result: 'cash',
            target: {
              id: '63d151eb-993a-4e80-9132-8107e9b92800',
              type: 'text',
            },
          },
          {
            result: 'default',
            target: {
              id: '8f81e367-6661-4be0-835e-7d598a3c9500',
              type: 'text',
            },
          },
        ],
      },
    },
```


## To document / Usage example

<!--
- How this is used?
- If possible, provide a snippet of code with a usage example.
-->

## Testing

The pull request...

- has unit tests
- has integration tests
- doesn't need tests because... **[provide a description]**