feedback knoweldege base with thumbsUp/thumbsDown (#2859)
## Description

Add feedback_knowledgebase event
Be able to add ThumbsUp and ThumbsDown buttons to give feedback on
webchat messages.
When the flow builder creates a message using the knowledge base Uses
the Text component with new withfeedback and inferenceid attributes

The withfeedback and inferenceid attributes cannot be declared using
camelCase because when they pass through the backen they come back with
everything in lowercase, it doesn't work with snake_case either. I'm not
sure but I think it's because of the html parser, the <message>
attributes have to be single words or use kebab-case.

## Context

Review this PR commit by commit:
[plugin-hubtype-analytics](dc8f37331340db098b60627766afe57a8f971fd6):
add feedback_knowledgebase event

[botonic-react](3fa97b0398873da8a084239a7a54696fdf0142e2): Refactor the
FooterMessage and add the FeedbackMessage which contains the ThumbsUp
ThumbsDown buttons and uses the trackEvent function when clicked.

[botonic-react](bb98f33a34fb2f181ac372a15d8bb8ca1ac9486f): Add the
possibility to inject the onTrackEvent function from the bot to the
webchat to create events from the frontend. Just like we already have
the onInit, onOpen, onClose, onMessage functions.

[plugin-flow-builder](8d90eb7f416ff7f20f57f6419445c693682b0295): When
the flow builder plugin uses the knowledge base to generate a message it
creates a @botonic/react Text with the new withfeedback and inferenceid
attributes needed for tracking.

The other commits are a refactor that renumbers the with_feedback
attrbute to feebackEnabled using camelCase

## Approach taken / Explain the design

Design
![Captura de pantalla 2024-06-21 a las 14 02
23](https://github.com/hubtype/botonic/assets/36898236/992cac28-fcdc-4109-bd9e-fc27d67e7204)

## To document / Usage example

How to pass the tracking function from bot to webchat
/webchat/index.ts
```typescript
export const webchat: WebchatArgs = {
  onInit: app => {
    window.botonicOnInit(app)
  },
  ...
  onTrackEvent: async (request, eventName, args) => {
    const pluginHubtypeAnalytics = new BotonicPluginHubtypeAnalytics()
    const htEventProps = {
      action: eventName as EventAction,
      ...args,
    } as HtEventProps
    await pluginHubtypeAnalytics.trackEvent(request, htEventProps)
  },
  theme: {
  ...
```

## Testing

Adds a test for the new feedback_knowledgebase event