botonic-react: remove getConfig function #BLT-1036 (#2902)
## Description

Remove the getConfig function that is run to get the response from the
botonic-config lambda.

## Context

In the new version of cli we will create the bot configuration at the
moment of deploying, it will be saved in the database and will be
updated in each deploy. We will not only have the plugins configuration.