plugin-flow-builder: replace bot action id for payload with params in pre funtion #BLT1187 (#2929)
## Description

Re-apply the old logic for buttons pointing to bot action nodes. 
When the user clicks on a button pointing to a bot action, the bot
receives in the payload the id of the bot action node. In the plugin's
pre function it checks if the id is from a bot action node and if so it
replaces the payload with the payload with the parameters defined in the
bot action node. This way, when reaching the bot routes, the payload is
already the expected one and it is not necessary to make a redirect to
run the core-bot again.

## Testing

Update test to check that the payload after pre function is the payload
with params defined in bot action node