botonic-react: pass Button code to typescript (#2936)
## Description

Review commit by commit. 
In the first commit only the file extension changes, in the following
ones you can see the applied changes.

## Context

Add the ClientInput because the input that is sent from the client
(webchat just like whatsapp, telegram etc) it is not necessary to put
the message_id or the bot_interaction_id. These values are set by the
backend

## Approach taken / Explain the design

Remove the bottom prop from the Button as it is not used anywhere and
the StyledButton used top and bootom but the top did not exist. The
button styles can be edited with a css stylesheet or by passing the
style in the Webchat theme.button.style.