dx-bundler-rspack: create a dx package to build with rspack #BLT-1197  (#2934)
## Description

Create a new dx package that has the dependencies to build a project
using as bundler rspack
It also contains a file rspack.config.ts. As explained in the README you
have to copy this file to the root of the botonic project.

## Context

In the README you can also see the scripts to use rspack to build the
project locally or to make a builder for production.