chore: update github action dependency #PLA-2658 (#2948)
<!-- _Set as [Draft
PR](https://github.blog/2019-02-14-introducing-draft-pull-requests/) if
it's not ready to be merged_.

[PR best practices
Reference](https://blog.codeminer42.com/on-writing-a-great-pull-request-37c60ce6f31d/)
-->

## Description

We have recieved the following email fron github:

`February 1st, 2025, we are closing down v1-v2 of actions/cache (read
more about it in this changelog announcement) as well as all previous
versions of the @actions/cache package in actions/toolkit. Attempting to
use a version of the @actions/cache package or actions/cache after the
announced deprecation date will result in a workflow failure. If you are
pinned to a specific version or SHA of the cache action, your workflows
will also fail after February 1st.`

For this reason we have to update all actions/cache v1 or v2
dependencies in all repos to v4 which we are using in other workflows.

<!--
- Must be clear and concise (2-3 lines).
- Don't make reviewers think. The description should explain what has
been implemented or what it's used for. If a pull request is not
descriptive, people will be lazy or not willing to spend much time on
it.
- Be explicit with the names (don't abbreviate and don't use acronyms
that can lead to misleading understanding).
- If you consider it appropriate, include the steps to try the new
features.
-->

## Context

<!--
- What problem is trying to solve this pull request?
- What are the reasons or business goals of this implementation?
- Can I provide visual resources or links to understand better the
situation?
-->

## Approach taken / Explain the design

<!--
- Explain what the code does.
- If it's a complex solution, try to provide a sketch.
-->

## To document / Usage example

<!--
- How this is used?
- If possible, provide a snippet of code with a usage example.
-->

## Testing

The pull request...

- has unit tests
- has integration tests
- doesn't need tests because... **[provide a description]**

---------

Co-authored-by: vanbasten17 <marcrabat@gmail.com>