botonic-react:  render carousel elements on whatsapp, using Multichannel #BLT-1395  (#2979)
## Description

Update the MultichannelCarousel to use Image and Text messages with
Buttons to render Carousel Elements in WhatsApp and other no webchat
channels.

If the Element has a URL button a WhatsApp message is created using the
WhatsAppCTAURLButton message component. If an Element has more than one
button and one of them is a URL button, two text messages will be sent,
one using WhatsAppCTAURLButton and the other using a Text message with
the rest of the Buttons. This way all the buttons can be rendered in
WhatsApp.

In a future PR I will refactor the Carousel to use typescript and ensure
that the elements always have Title

## Context

Until now for the Carousels in WhatsApp only the text of the button was
rendered with a number in front for the older bots, those who used
contentful as cms and WhatsApp without buttons. **This will no longer
work, new bots should use WhastApp with buttons.**

## Testing

The pull request...

Update tests and update snapshots