plugin-flow-builder: nluResolution available in first interaction (#3004)
## Description

In order to have available the detection of keywords, smart-intents and
intents-babel in the first interaction, as long as the user input is of
text type and there is no payload, a match with keywords, smart-intents
or intents-babel will be searched. This information is stored in the
request.inpu.nluResolution together with the payload pointing to the
matched node.

In this way it is no longer necessary to look for the match within the
action when it is the first interaction, since the necessary payload
will be stored in the input.nluResolution.payload.


## Testing

Update tests