botonic-react: ReactBot class with typescript (#3003)
## Description

Pass ReactBot class to typescript
Add RequestCoreContext type. This is the request that the CoreBot passes
to the actions and is used from the RequestContext.
Create requestCoreContext with the values and functions that need
RequestContext inside actions
Not only use this requestCoreContext for actions, use it for pre and
post functions od the plugins and also in routes