botonic-react: refactor webchat theme and create default theme #BLT-1472 (#2991)
## Description

Use typescript for the `WebchatTheme`.
Add a `ThemeProvider` of styled-components to avoid having to pass as
parameters the theme attributes to all the components.
Define a `defaultTheme` for the webchat so that if nothing is defined in
the bot it still has the same styles. This object contains the default
values that were previously in a constant (WEBCHAT.DEFAULT)

Review commit by com
mit for more context

**IMPORTANT**
Remove `persistentMenu`, `coverComponent`, `blockInputs`,
`enableEmojiPicker`, `enableAttachments`, `enableUserInput` and
`enableAnimations` from `WebchatProps` and pass it inside `theme`. This
values can be defined outside the theme in `webchat/index` file in bot
code (the attributes defined in this file are the WebchatArgs used by
WebchatApp)
In a future refactor it would be better to only be able to define these
attributes in one way.

With this changes in a future PR we can remove the getThemeProperty
function and use the webchatState.theme to retrieve all the attributes
of the theme so they will always have the type they really have.

## Context

To be able to use the theme attributes in custom components written
inside the bot code without having to pass the attributes as parameters
we need to add the file theme.d.ts in the webchat folder.

webchat/theme.d.ts
```typescript
import 'styled-components'

import { WebchatTheme } from '@botonic/react'

// With this declaration, we can use the theme prop in styled components with TypeScript defined in WebchatTheme
declare module 'styled-components' {
  export interface DefaultTheme extends WebchatTheme {}
}

```