plugin-flow-builder: use new user fields and default_locale_code #BLT-1581 (#3008)
## Description

Use new user locale, country and system_locale from request.session.user
Use default_locale_code when not be able to resolve the locale to
generate the contents

## Context

The plugin no need more the getLocale function. First use the new user.
system_locale field to try to show contents with the locale of the user.
For example if the system_locale is `es-ES` first find if this locale is
available in locales of the flow, if this is not available try to find
`es`, if this is not available resolve with the default_locale_code from
the flow.