Some fixes before 0.35.0 (#3014)
## Description

In this PR there are changes in several plugins that are necessary for
the next version (0.35.0)

**@botonic/plugin-hubtype-analytics**
- Uses BotContext instead of BotRequest
- Remove the optional functions that could be passed in through the
constructor to obtain the locale and the country

**@botonic/react**
- Creates a hook for the feedback tracking function of a knowledge base
message. Inside this hook the request is recreated as a BotContext (in
the frontend there is no BotContext) which is what the hubtype analytics
plugin needs to receive.

**@botonic/plugin-flow-builder**
- Removes the getLocale function that was passed through the constructor

**@botonic/plugin-knowledge-bases**
- use BotContext instead of BotRequest

**@botonic/plugin-google-translation**
- use getUserLocale instead of __locale