feat(plugin-flow-builder): always adds match contents with some nlu  (#3018)
## Description

plugin-flow-builder always adds match contents with some nlu when bot
pass from conversation start node. Until now it only added contents of a
match with NLU when it was the first interaction.

## Tests

Update basicFlow, now smart intents titles are in snake_case.
Add a test to check that executeConversationStart always add match of
keywords or smart intents in all interactions.