botonic-core: infer type from Route to matcher (#3027)
## Description

To be able to infer the type of the plugins and the user.extra_data to
the route matchers, it is necessary to infer the type first to Route


## To document / Usage example

Example of how to infer the type through the Route. 
With this change you have de request.plugins with the type of BotPlugins
and request.session.user.extra_data with the type of UserData inside the
matchers.

```typescript
import { Route } from '@botonic/core'
import { UserData } from './domain/user-data'
import { BotPlugins } from './plugins'
import { BotRequest } from './types'

export function routes(request: BotRequest): Route<BotPlugins, UserData>[] {

  return [
    {
      path: 'welcome-action',
      payload: WELCOME_PAYLOAD,
      request: (request: BotRequest) => {
        console.log('WELCOME_ACTION', request.plugins)
        return request.session.is_first_interaction
      },
      action: WelcomeAction,
    },
...
```