plugin-flow-builder: add contentFilters #BLT-1683 (#3029)
## Description

Be able to add contentFilters in the plugin's settings.
These filters will be applied to the contents just before tracking.

## Context
The AiAgent feature is unfinished, for now I skip the tests and the flow
builder constructor will set undefined so the AiAgent will never work.
When we finish the feature I will activate it again.

## Testing

Add a test to check that a button can be removed with a filter