plugin-ai-agents:  use langGraph to manage agents and tools #BLT-1666  (#3033)
## Description

This plugin is used to activate an AIAgent in the bot. It uses langGrah
to create the model, agent and tools.
For now only Azure openAI models are supported. 
When deploying the bot it is necessary to have the environment
variables:
- AZURE_OPENAI_API_KEY
- AZURE_OPENAI_API_DEPLOYMENT_NAME
- AZURE_OPENAI_API_VERSION
- AZURE_OPENAI_API_BASE

The plugin adds by default some tools to stop the agent from continuing
the conversation:
- handoffToHumanAgent
- outOfContext
- finishConversation

Updates the flow-builder plugin that receives the agent's response.
Change the way the exit tool is handled so that the AIAgent does not
respond.

---------

Co-authored-by: Eric Lozano <ericlozano98@gmail.com>