botonic-cli: get tools, payloads and webviews from bot to save in bot config in every deploy #BLT-1665 (#3035)
## Description

We need to save in the backend along with the botConfig the tools,
payloads and webviews that the bot can use.

## Context

To achieve this, the bot must have a file in src/bot-config.ts that
exports a botConfig object that follows the following interface:

```typescript
interface Tool {
  name: string
  description: string
}

interface Webview {
  name: string
}

interface BotConfigJSON {
  tools: Tool[]
  payloads: string[]
  webviews: Webview[]
}
````

## To document / Usage example

In the bot we can have a file in `src/server/tools.ts` where the
customTools are defined, it is an array of Tools (this interface comes
from the plugin-ai-agents.
```typescript
export customTools: CustomTool[] = [{
    name: 'get_date',
    description: 'Get date',
    schema: z.object({}),
    func: async () => {
      const date = new Date()
      console.log('get_date', { date })
      return date.toISOString()
    },
  }]
```

This array will be used to create the plugin-ai-agents configuration in
`src/server/config`
```typescript
function getAiAgentConfig(): PluginAiAgentOptions {
  return {
    customTools: customTools,
  }
}
```

This array will be used with a map to create botConfig.tools in
`src/bot-config.ts`
```typescript
import { RATING_PAYLOAD } from './server/constants'
import { customTools } from './server/tools'

const aiAgentTools =
  customTools?.map(customTool => ({
    name: customTool.name,
    description: customTool.description,
  })) || []

export const botConfig: BotConfigJSON = {
  tools: aiAgentTools,
  payloads: [RATING_PAYLOAD],
  webviews: [
    { name: 'ExampleWebview' },
    {
      name: 'FlowBuilderWebview',
    },
  ],
}
```

## Testing

Update deploy test

---------

Co-authored-by: Eric Lozano <ericlozano98@gmail.com>